requirements.txt
# Core dependencies
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Query Engine
duckdb==0.9.2

# Redis & Caching
redis==5.0.1
celery==5.3.4

# API
fastapi==0.105.0
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6

# Web UI
streamlit==1.29.0

# Carbon & Energy APIs
requests==2.31.0
httpx==0.25.2

# System Monitoring (with proper build)
psutil==5.9.6

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Scheduling
apscheduler==3.10.4
croniter==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
slowapi==0.1.9

plotly==5.18.0
